custom_requirements.txt
youtube_transcript_api
langdetect
absl-py==2.2.2
ai4bharat-transliteration==1.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aksharamukha==1.9.7
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==4.9.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
bangla==0.0.5
bitarray==3.4.1
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cached-property==2.0.1
catalogue==2.0.10
cdifflib==1.2.9
cffi==1.17.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==45.0.2
cycler==0.12.1
dataclasses==0.6
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
editdistance==0.8.1
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
exejs==0.0.4
fairseq==0.12.2
ffmpeg-python==0.2.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
future==1.0.0
g2pkk==0.1.2
gevent==25.5.1
greenlet==3.2.2
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hydra-core==1.0.7
inflect==7.5.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jh2==5.0.9
jieba==0.42.1
jsonlines==1.2.0
julius==0.2.7
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
lightning-utilities==0.14.3
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
mir_eval==0.8.2
mock==5.2.0
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.18
murmurhash==1.0.12
nemo_text_processing==1.1.0
niquests==3.14.0
nltk==3.9.1
num2words==0.5.14
numba==0.60.0



onnxruntime==1.22.0
pathos==0.3.4
pb-bss-eval==0.0.2
pesq==0.0.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pox==0.3.6
ppft==1.7.7
preshed==3.0.9
progressbar2==4.5.0
propcache==0.3.1
protobuf==3.20.0
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydload==1.0.9
pyenchant==3.2.2
pykakasi==2.3.0

pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pyreadline3==3.5.4
pysbd==0.3.4
pystoi==0.4.1
python-crfsuite==0.9.11
python-utils==3.9.1
pytorch-lightning==2.5.1.post0
pytorch-ranger==0.1.1
qh3==1.5.1
rich==14.0.0
scikit-learn==1.6.1
scipy==1.11.4
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250129
SudachiPy==0.6.10
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.3.4
threadpoolctl==3.6.0
torch-optimizer==0.1.0
torch-stoi==0.2.3
torchaudio==2.7.0
torchmetrics==0.11.4
trainer==0.0.36
translators==5.9.9
TTS==0.22.0
typeguard==4.4.2
typer==0.15.4
typing-inspection==0.4.0
tzlocal==5.3.1
ujson==5.10.0
umap-learn==0.5.7
urduhack==0.1.4
urllib3-future==2.12.921
wasabi==1.1.3
wassima==1.2.2
weasel==0.4.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
yarl==1.20.0
zope.event==5.0
zope.interface==7.2

indic_numtowords==1.0.2

# omegaconf==2.0.6
omegaconf>=2.0.5,<2.1

# pynini==2.1.6.post1   






# conda install -c conda-forge pynini=2.1.6


# pip install -r C:\Users\ROG\Desktop\mafatlaal\custom_requirements.txt